Adds uses to all Wagtail models.

+++ wagtail/snippets/views/snippets.py
@@ -10,6 +10,7 @@ from wagtail.admin import messages
 from wagtail.admin.edit_handlers import ObjectList, extract_panel_definitions_from_model_class
 from wagtail.admin.forms import SearchForm
 from wagtail.admin.utils import permission_denied
+from wagtail.core.collectors import get_paginated_uses
 from wagtail.search.backends import get_search_backend
 from wagtail.search.index import class_is_indexed
 from wagtail.snippets.models import get_snippet_models
@@ -230,7 +231,9 @@ def delete(request, app_label, model_name, pk):
 
     instance = get_object_or_404(model, pk=unquote(pk))
 
-    if request.method == 'POST':
+    uses = get_paginated_uses(request, instance)
+
+    if request.method == 'POST' and not uses.are_protected:
         instance.delete()
         messages.success(
             request,
@@ -244,6 +247,7 @@ def delete(request, app_label, model_name, pk):
     return render(request, 'wagtailsnippets/snippets/confirm_delete.html', {
         'model_opts': model._meta,
         'instance': instance,
+        'uses': uses,
     })
